+++ x/slashing/module.go
@@ -12,6 +12,9 @@ import (
 	store "cosmossdk.io/core/store"
 	"cosmossdk.io/depinject"
 	govtypes "cosmossdk.io/x/gov/types"
+	"cosmossdk.io/x/slashing/keeper"
+	"cosmossdk.io/x/slashing/simulation"
+	"cosmossdk.io/x/slashing/types"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -19,9 +22,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/cosmos/cosmos-sdk/x/slashing/keeper"
-	"github.com/cosmos/cosmos-sdk/x/slashing/simulation"
-	"github.com/cosmos/cosmos-sdk/x/slashing/types"
 	staking "github.com/cosmos/cosmos-sdk/x/staking/types"
 )